# Add continuous benchmarking setup