--- a/.github/workflows/python-lint-tests.yml
+++ b/.github/workflows/python-lint-tests.yml
@@ -87,14 +87,16 @@ jobs:
     # PYTEST STRATEGY
     # macOS is the only runner who has working ModernGL behaviour
     # macOS + 3.8 is used for code coverage
-    - name: Pytest (code coverage)
-      run: poetry run pytest --mpl --mpl-baseline-path tests/data/baseline --cov=./ --cov-report=xml
+    - name: Pytest (code coverage + benchmark)
+      run: |
+        poetry run pytest --cov=./ --cov-report=xml
+        poetry run pytest tests/test_benchmarks.py --benchmark-enable --benchmark-json output.json
       if: matrix.os == 'macos-latest' && matrix.python-version == '3.8'
     - name: Pytest
-      run: poetry run pytest --mpl --mpl-baseline-path tests/data/baseline
+      run: poetry run pytest
       if: matrix.os == 'macos-latest' && matrix.python-version != '3.8'
     - name: Pytest (no image similarity check)
-      run: poetry run pytest --mpl --mpl-baseline-path tests/data/baseline --skip-image-similarity
+      run: poetry run pytest --skip-image-similarity
       if: matrix.os != 'macos-latest'
     - name: Upload test results
       if: always()
@@ -108,6 +110,20 @@ jobs:
         token: ${{ secrets.CODECOV_TOKEN }}
         file: ./coverage.xml
       if: matrix.os == 'macos-latest' && matrix.python-version == '3.8'
+    - name: Store benchmark result
+      uses: rhysd/github-action-benchmark@v1
+      if: matrix.os == 'macos-latest' && matrix.python-version == '3.8'
+      with:
+        tool: 'pytest'
+        output-file-path: output.json
+        # push to gh-page
+        github-token: ${{ secrets.PERSONAL_GITHUB_TOKEN }}
+        auto-push: true
+        # Show alert with commit comment on detecting possible performance regression
+        alert-threshold: '150%'
+        comment-on-alert: true
+        fail-on-alert: true
+
 
   #################
   # TESTS WINDOWS #
@@ -138,7 +154,7 @@ jobs:
       run: poetry install
     - name: Pytest
       run: |
-        poetry run pytest --mpl --mpl-baseline-path tests/data/baseline --skip-image-similarity
+        poetry run pytest --skip-image-similarity
     - name: Upload test results
       if: always()
       uses: actions/upload-artifact@v2