--- a/pyproject.toml
+++ b/pyproject.toml
@@ -60,6 +60,7 @@ recommonmark = { version = ">=0.6,<0.8", optional = true }
 coverage = {extras = ["toml"], version = "^5.4"}
 pytest = "^6.2.1"
 pytest-cov = "^2.11.0"
+pytest-benchmark = "^3.2.3"
 black = "^20.8b1"
 isort = "^5.7.0"
 mypy = "^0.800"
@@ -74,6 +75,8 @@ docs = ["Sphinx", "sphinx-click", "sphinx-autodoc-typehints", "sphinx-rtd-theme"
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
+[tool.pytest.ini_options]
+addopts = "--mpl --mpl-baseline-path tests/data/baseline --benchmark-disable"
 
 [tool.coverage.run]
 branch = true