… poetry.lock
… tests/data/test_svg/benchmark/300_beziers.svg
… tests/data/test_svg/benchmark/500_circles.svg
… tests/data/test_svg/benchmark/500_polylines.svg
… tests/data/test_svg/benchmark/7k_lines.svg
… tests/data/test_svg/benchmark/900_quad_beziers.svg
… tests/data/test_svg/benchmark/bar_nodef_path_polylines.svg
… tests/data/test_svg/benchmark/multi_skull.svg