@abey79 2021-01-28
Deferring `_build()` in `vpype_viewer.Engine` yielded the following improvement:

```
Before optimizing:

-------------------------------------------------------------------------------------------------- benchmark: 5 tests --------------------------------------------------------------------------------------------------
Name (time in ms)                                  Min                   Max                  Mean             StdDev                Median                IQR            Outliers     OPS            Rounds  Iterations
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_benchmark_viewer[preview]                954.8367 (1.0)      1,001.5650 (1.0)        979.1492 (1.0)      18.6876 (1.0)        978.8370 (1.0)      30.0780 (1.0)           2;0  1.0213 (1.0)           5           1
test_benchmark_viewer[outline]              1,546.8146 (1.62)     1,679.9941 (1.68)     1,638.3394 (1.67)     53.1831 (2.85)     1,656.8107 (1.69)     53.6346 (1.78)          1;0  0.6104 (0.60)          5           1
test_benchmark_viewer[outline_colorful]     1,705.3878 (1.79)     1,808.9265 (1.81)     1,760.6042 (1.80)     40.6611 (2.18)     1,760.0272 (1.80)     62.6450 (2.08)          2;0  0.5680 (0.56)          5           1
test_benchmark_viewer[outline_pen_up]       1,807.2741 (1.89)     1,984.7084 (1.98)     1,874.0171 (1.91)     68.3519 (3.66)     1,869.6960 (1.91)     80.7459 (2.68)          1;0  0.5336 (0.52)          5           1
test_benchmark_viewer[outline_points]       1,920.9737 (2.01)     2,124.0354 (2.12)     2,047.9244 (2.09)     78.7294 (4.21)     2,068.5779 (2.11)     98.2199 (3.27)          1;0  0.4883 (0.48)          5           1
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

After deferring _rebuild():

---------------------------------------------------------------------------------------------- benchmark: 5 tests ----------------------------------------------------------------------------------------------
Name (time in ms)                                Min                 Max                Mean             StdDev              Median                IQR            Outliers     OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_benchmark_viewer[preview]              352.4981 (1.0)      458.2220 (1.0)      400.4586 (1.0)      41.7608 (2.21)     399.3118 (1.0)      64.6774 (3.44)          2;0  2.4971 (1.0)           5           1
test_benchmark_viewer[outline]              383.6201 (1.09)     483.0642 (1.05)     441.1018 (1.10)     42.5857 (2.25)     435.6457 (1.09)     71.5712 (3.81)          1;0  2.2671 (0.91)          5           1
test_benchmark_viewer[outline_colorful]     422.6880 (1.20)     525.1907 (1.15)     459.9273 (1.15)     41.1411 (2.17)     440.1814 (1.10)     53.6903 (2.86)          1;0  2.1743 (0.87)          5           1
test_benchmark_viewer[outline_pen_up]       470.6696 (1.34)     593.6541 (1.30)     522.2302 (1.30)     51.7979 (2.74)     504.0013 (1.26)     85.2293 (4.54)          1;0  1.9149 (0.77)          5           1
test_benchmark_viewer[outline_points]       495.0083 (1.40)     546.2495 (1.19)     515.5171 (1.29)     18.9354 (1.0)      510.8274 (1.28)     18.7916 (1.0)           2;0  1.9398 (0.78)          5           1
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```